rely on garbage collector to manage user-supplied node data
This prevents segfaults from double free() calls when using
Patricia#destroy on node data or the same object twice.
Patricia#destroy and Patricia#clear are now no-ops for backwards
compatibility.

One (hopefully welcome) side effect of this change is that
arbitrary Ruby objects may be attached to nodes, not just
strings.  String objects will continue to alway be duplicated
when returned to the user for backwards compatibility, but
arbitrary Ruby objects may be attached to each node.  This
feature can now be used to track node usage data directly in the
Patricia object without the need for a separate lookup table.

The only downside of this change is that the GC mark phase
takes a tiny bit longer due to the presence of more Ruby
objects.

There are also additional tests for the new behavior, including
an expensive test to force GC pressure and ensure there are no
memory leaks.