enable Patricia#dup to work
Instead of defining the Patricia.new singleton method, we
only need to define the allocation function.  The allocation
function allows "dup" and "clone" methods to be called on
Patricia instances and we will walk the original tree
to copy node prefixes and data into the new one.

This also allows us to subclass Patricia and create
objects from that subclass.