remove prefix_toa() function and ensure thread-safety
prefix_toa() is not used anywhere outside of debugging, so
remove that since this is well-tested code.

The destination buffer of prefix_toa2x() is now enforced for the
caller.  The previous array of static buffer is both wasteful
and not safe since the i++ used to switch buffers is never
guaranteed to be atomic.